Fix D2 Name
